+++ src/shared/ui/input/select/Select.tsx
@@ -1,7 +1,7 @@
 import * as S from './Select.styles';
 import { useRef, useState } from 'react';
 import SelectDownIcon from '../../assets/icons/ic_select.svg';
-import useOutsideClick from '../../lib/useOutsideClick';
+import useOutsideClick from '../../../lib/useOutsideClick';
 
 interface SelectProps {
   options: string[];
